Fixed compile warnings. Fixed line endings on modified files.

+++ src/SOIL.c
@@ -87,7 +87,7 @@ unsigned int SOIL_direct_load_DDS(
 		int loading_as_cubemap );
 unsigned int SOIL_direct_load_DDS_from_memory(
 		const unsigned char *const buffer,
-		int buffer_length,
+		unsigned long buffer_length,
 		unsigned int reuse_texture_ID,
 		int flags,
 		int loading_as_cubemap );
@@ -1531,7 +1531,7 @@ const char*
 
 unsigned int SOIL_direct_load_DDS_from_memory(
 		const unsigned char *const buffer,
-		int buffer_length,
+		unsigned long buffer_length,
 		unsigned int reuse_texture_ID,
 		int flags,
 		int loading_as_cubemap )
